WT-14568 Compat tests: handle absence of stable toolchain file (#11913)
Recently a toolchain file for compatibility tests was switched from
`mongodbtoolchain_v4_gcc.cmake` to `mongodbtoolchain_stable_gcc.cmake`.
Later it turned out that some branches that are randomly picked for a
compatibility testing don't contain this file.

This patch fixes it by using the stable toolchain only if it exists.