Add evergreen integration
Add a --generate-tests command that is called by evergreen. This creates a file that is passed into generate.tests and will create testing tasks for the relevant components in evergreen

Example command:
`python3 smart_test.py --generate-tests foo`

`foo` is the name of the buildvariant. We need it so we know which variant to add the tests to