WT-14039 avoid double eviction with scrub eviction (#11963)
This change updates the code for scrub eviction, to avoid double
evicting a page when the cache is under pressure. It is currently
possible for a dirty page to be scrub evicted and read back in the
cache, then clean evicted afterwards. To avoid this, a condition has
been added to skip scrub eviction unless either clean eviction threshold
has not been reached, or the page has high read generation. If we're
over the clean eviction threshold and the page's read generation is low,
we just evict the page.