WT-13703 Truncate history store for a dropped file (#11959)
This PR is to fix the bug where a newly created table having same tree
id as a file dropped previously, causing potential data corruption as
history store may still contain records for the dropped file.

The PR also includes a refactor which moves history file truncate
function `__wt_hs_btree_truncate` to `hs_cursor.c`.

Note this is a problem only for standalone versions, as mongodb will not
drop a table until the oldest timestamp past, so when a table is dropped
all the records for that table in history store are earlier than the
oldest timestamp and they will never be read.