WT-14598 Add disaggregated storage block manager (#11960)
This is the second in a series of disagg storage PRs.

There aren't any tests for this since the disagg tests will only make
sense after some more components have been merged. This is also the
reasoning behind adding a bunch of stuff to exception lists, e.g. for
s_all.

The best place to start reviewing is `src/include/block.h`. That has
some interface changes, and the rest of the changes flow from there.

I'll be reviewing this too, mostly to convert comments into tickets with
a `FIXME-WT-xxxxx`. I do welcome comments and improvement suggestions,
but for the sake of not getting bogged down (there is a lot of code
still left to merge) I do plan to merge this even if not all
conversations are resolved.