WT-14533 Skip ingest table lookups on primary (#11972)
(Pushing this to the old branch so we can vendor it directly)

This isn't just easy "gate the existing code with an `if`" option, I
took the opportunity to refactor it a bit. The summary is that I moved
the `set_key`+`search` bit out to its own function, and that enabled a
bunch of follow-on fixes.

T2 claims this improves read throughput by about 4%. That seems high,
but I'm not sure what the noise floor is. However it definitely reduces
the number of cache accesses per read from 9 to 8.