WT-14039 avoid double eviction from scrub eviction (#11976)
This change updates the code for scrub eviction, to avoid double
evicting a page when the cache is under pressure. It is currently
possible for a dirty page to be scrub evicted and read back in the
cache, then clean evicted afterwards. To avoid this, a condition has
been added to skip scrub eviction unless either clean eviction threshold
has not been reached, or the page has high read generation. If we're
over the clean eviction threshold and the page's read generation is low,
we just evict the page.

This was previously reverted in #11974, as the python test
intermittently failed. This could be due to race conditions when trying
to access the stats cursor before the stats are updated. A timeout and
retry loop has been added to the test to avoid this. The settings have
been updated to be less aggressive on eviction to prevent background
threads disrupting the test.