WT-14649 track cache fill ratio stats whenever app thread eviction requested (#11981)
This PR adds additional stats and tracks the percentage of cache that's
filled up whenever application threads are pulled into eviction. We
measure whether the cache is < 25%, >= 25% to < 50%, >= 50% to < 75%, or
>= 75% full when application threads are used in eviction. We can use
these stats to determine if we can improve eviction by taking into
account the amount of cache usage before using app threads in eviction.