WT-14676 Change eviction stats tests to be more consistent (#11989)
The tests for eviction trigger statistics are non-deterministic as it's
hard to control which pages get evicted. We can change the test to check
that any eviction happens and increments related stats instead of
checking that specific kinds of eviction happens. This should make the
test more consistent and still check that eviction related stats are
incrementing as they should be when eviction thresholds are met.