WT-14270 Fix modifications vector reconstruction (#11990)
Fix for the memory leak that was detected during format testing.

While trying to read an overflow page with `WT_UPDATE_MODIFY` enabled
there can be a race with checkpoint reconciliation removing the overflow
items, so we need to retry collecting modifications since reconciliation
would store a complete update to the update list.

This PR properly frees the memory of the modification vector before
retrying.