WT-13694 Ensure truncate skips unseen keys in the model (#11985)
In WT, truncate skips over both newly created uncommitted keys and any
(new) keys that are invisible to the truncate transaction. That said, in
the model, we don't perform a search or search_near when truncating. As
a result, keys that were newly created, or newly committed incorrectly
gets included as part of the truncation. This, of course, is not how WT
behaves.

As part of this PR, (1) the model is to explicitly check whether the key
exists for the truncate transaction. (2) Added tests