WT-14753 Check a filename argument content before opening a block manager (#12010)
This change was inspired by a coverity report which highlighted that
passing NULL as the `filename` argument could lead to a NULL
dereference. While it is not really dangerous since it would simply fail
later, it's still worth adding an assertion to make the reason of
failure clearer and more meaningful.