WT-14619 Merge layered tables (#11998)
This is the third in a series of disagg storage PRs.

Again, the tests don't quite make sense yet so mostly aren't in this PR.

The best place to start reading here are `src/conn/conn_layered.c` and
`src/cursor/cur_layered.c`.

This also includes a bunch of other supporting changes. Some of those
changes are in the block manager, so I've added Yury as a reviewer but
there isn't a huge amount to review there.

I'd also hope that we can do another async review where we raise tickets
for issues even if this gets merged.