WT-12542 Avoid retry for next and prev API calls (#12021)
Retry logic is in place for cursor search, search_near, next, and prev
API calls if they receive a WT_ROLLBACK error. It's OK for search and
search_near to retry (as they are idempotent), but not for next and prev
as when the btree layer returns a WT_ROLLBACK error it resets the
cursor. A retry without the cursor positioned is logically incorrect.

This PR crafts a new Python test trying to simulate the described
errorous scenario, the test will hung doing retry repeatedly for a
cursor_next() call, and pass with the retry removed. The fix is to
remove the retry for the curosr next and prev API calls.