WT-14793 Track metrics when we are looping in evict_app_assist_worker (#12037)
When we are in __wti_evict_app_assist_worker, we do not report cache
metrics while it is stuck for x amount of time, and we report the total
once this exits. The problem is, if we are not updating while it is
stuck, our metrics are unable to tell we are in a cache stuck state to
alleviate the cache pressure (cache pressure thread in server) and we
won't be able to alert our thread until after it exits.

We want to update the metrics within the [while
loop](https://github.com/wiredtiger/wiredtiger/blob/develop/src/evict/evict_lru.c#L2897-L2965)
where it can get stuck.

We have FTDC data to show that these metrics make a difference. 

We don't double count cause we update the time_start.