WT-14461 Fix workgen test failures on machine ARMV9 using V5 toolchain (#11988)
The workgen python tests import workgen and wiredtiger shared library
which has direct dependency to the libstdc++ library. Digging into a
rabbit hole found that the wiredtiger.so library has a dependency on
snappy library which has a dependency on the libstdc++ library. For some
reason importing the libstdc++ from snappy causes a segmentation fault
within workgen python tests.

~~We have found out that the ordering here matters. This is an odd case
and is not a common use case of the wiredtiger.so library. The final
discussed solution is to import workgen first and then the wiredtiger.so
library. In that way the correct libstdc++ can be symbolised, and will
not cause further problems in workgen.~~
From further investigation we were able to find that libstdc++ symbols
are colliding. As a solution hide all public symbols from workgen such
that it doesn't collide with snappy library.