WT-14804 Fix crash when printing an invalid dhandle's name on sweep walk (#12038)
A bug was introduced in commit
https://github.com/wiredtiger/wiredtiger/commit/aa8b5a7c0c7cd13d5c08ce5c3566ee1ee9078864
, in `__sweep_file_dhandle_check_and_reset_tod` function which calls
`__wt_conn_dhandle_find`. Inside `__wt_conn_dhandle_find()` we find a
dhandle and assign the `session->dhandle` to it.
Later inside the sweep walk, `__sweep_check_session_callback` is called,
but the dhandle might already been deleted but `session->dhandle` still
points to that handle, when it tries to print the session's dhandle
name, it will point to an invalid pointer.
This PR fixes it by assigning back the session's dhandle to null after
calling `__wt_conn_dhandle_find()` to avoid this bug from happening.