WT-14858 Forbid to prepare a transaction before the stable timestamp if the preserve_prepare config is on (#12073)
When the preserve_prepare config is enabled, we cannot prepare a
transaction older than or equal to the stable timestamp. Otherwise, we
cannot ensure we will write a complete transaction to the checkpoint.
Throw an error in this case.

The change doesn't impact the existing versions as it is behind a
feature flag. A python test is added to validate the above case.