WT-14833 Fix ENABLE_TCMALLOC propagation (#12060)
It turned out that `ENABLE_TCMALLOC` was incorrectly set on some
platforms since the only way to set it for `PREPARE_TEST_ENV` is to
define it as an `expansions` for a variant. And if a platform define
task specific `ENABLE_TCMALLOC` it cannot be propagated to
`PREPARE_TEST_ENV`. So, the only way to enable/disable TCMalloc
correctly for some tests is to create separate variants, and this is
what this PR does.

It also does two changes that I decided not to move to separate PRs:
- Fail the test if ENABLE_TCMALLOC is set, but TCMalloc library does
exist on the system
- Rename `*-sanitizer-*` tests to `*-asan-*` since all these tests
enables ASAN and "sanitizer" sounds ambiguous from my point of view
(also we already have other "asan" tests so it's for consistency too)