WT-14727 Update workgen latency metrics to output bucket count (#12000)
At the moment the latency output metrics contains the avg, min and max
along with the plot of the latencies.
It will be better to understand the count of the latency buckets
corresponding to us, ms and secs.
By default only print the latency bucket count, instead of the latency
bucket plot