WT-14696 Merge precise checkpoints (#12090)
NOTE: the vast majority of this code has already been reviewed on its
way into the disagg branch. That said, this branch does serve as a good
opportunity to step back and take a look at the feature more broadly.
That review can/should continue even after this is merged.

There is some (trivial) new code -- the macro for `WT_ACQUIRE_READ` had
a spurious extra semicolon, props to @ajmorton for spotting that.