WT-14648 Fix log subsystem returning EBUSY from conn->close() (#12050)
The investigation identified a real scenario where EBUSY could return
from conn->close(). The scenario is as such:
1. Log server is doing a log_printf() which requires a slot
2. Connection->close() needs to flush it's lsn and requires a slot to do
so
3. Connection->close() can't access the slot and returns EBUSY

This tickets makes sure that conn->close() does not return EBUSY
interacting with the log subsystem. The reproducer failing is on the
ticket.

**Update:** During this investigation i found two race condition bugs
wrlsn due to conn->close() shutting down. More informatino can be found
on the ticket.