WT-14832 Add workload get to test/model operation and improve prepare semantics (#12048)
This change is part of my work on extending model based validation of
WiredTiger. I needed to extend the workload to include `get` operations,
and it exposed a mismatch between the model and WT in terms of prepare
semantics.

This PR includes:
- new `get` workload op, which becomes part of the generated /
reproduced workloads the model can handle
- tests that expose a relationship between get operation outcomes (not
found, prepare conflict, ok) and interleavings with prepared
transactions.
- changes to how `get` operations are modeled, as well as whether
prepared transactions are visible to newly created transactions. I
commented the reasoning for each new condition I had to add to the `get`
operation model.