WT-14746 Fix reserve operation when using layered cursor (#12069)
The assertion inside cursor_lookup() function requires the
WT_CLAYERED_OPEN_READ flag set. From the stacktrace, test/format is
trying to perform a cursor->reserve() operation and hits this assert.
The cursor->reserve() reserve an existing record so a subsequent write
is less likely to fail due to a conflict between concurrent operations
(ref:
[reference](https://source.wiredtiger.com/develop/struct_w_t___c_u_r_s_o_r.html#a3acc1fd180c4a4220b459734305c90a6).

Internally WiredTiger treats reserve() operations as an update operation
without the overwrite flag set and functionally just needs to position
the cursor. Inside disaggregated storage world, we perform a
cursor_lookup and the reserve() operation is considered an update
operation. The configuration of the layered cursor is done inside
cursor->enter().

The cursor->enter() doesn't handle the reserve operation. Therefore
there are two required fixes for this change:
1. Unset the overwrite flag when performing reserve() operation
2. Update the cursor_enter() function to accomodate for this change