WT-14947 Suppress perf critical warnings (#12107)
My assumption is that the recent merges has slowed down the system to
the point where the example would perform evictions or need to fetch a
page from disk. This causes the new warnings.

All related warnings are under eviction or row search. Both of these are
considered perf critical warnings and need deeper investigation.