WT-14953 Fix layered table manager initialization (#12112)
Currently, the layered table manager is initialized during the first
table opening. However, some parts of the code that are unrelated to
tables—for example, reconfiguration—might rely on the manager’s fields
before it is even initialized.

To address this, we decided to move the manager initialization routine
to __wti_disagg_conn_config(), which is executed as part of
wiredtiger_open(). This also eliminates the need to synchronize manager
initialization across multiple threads.