WT-12337 Add runtime checks instead of wt_assert in packing_inline (#12126)
Some asserts in packing_inline.h can cause the whole program to crash
(if WT_ASSERT is enabled) or ignore an error (if WT_ASSERT is disabled).
This replaces all uses of WT_ASSERT with a check that returns an error
code on a fail. There is a function, `__wt_struct_size_adjust` that has
no return value. I left the WT_ASSERT in that function for now, but we
may have to change the function signature if we want it to return an
error code.