WT-14971 Fix broken cursor function entry macro call (#12116)
This changes an argument that is only "stringified" and is used to set a
session name while the session is inside an API call. So basically, the
last argument becomes part of some output... like
`"WT_CURSOR.cbt->dhandle"`, which is confusing, we want the output to be
`"WT_CURSOR.insert"` See all the other uses of
`CURSOR_UPDATE_API_CALL_BTREE` in this function for correct usage.

Co-authored-by: Don Anderson <donald.anderson@mongodb.com>