WT-14969 Skip including a prepared update in the delta if it has been written (#12114)
This change skips a prepared update in the delta if it has already been
written. It also include the change to not include a delete in delta if
it has been durable.

Summarize the reason behind this change (this might be the problem
you're solving, or the context around the request) and the solution you
have chosen.