WT-14987 Fix missing endianness conversion before disagg block manager write (#12130)
`__bmd_checkpoint_pack_raw` forgot to do an endianness conversion.

`__wti_block_disagg_write_internal` might seem like a better place to
fix this, but it doesn't know what type of data it's writing, it just
gets a `WT_ITEM`.

Also note that this does NOT solve every possible endianness issue --
this is the smallest possible change to get the CI-blocking crashes
fixed.