WT-14609 Revert block cache compression calculation (#12145)
This was something of a known issue previously. Ideally we'd have
something
a bit fancier that knows how to apply a different heuristic based on
whether
we can generate deltas (i.e. in disagg), but since the release process
is
kicking off again, this is technically a release blocker. So just do the
simple thing and revert this code to its pre-disagg state.

This doesn't warrant a `FIXME`, since it's not technically broken --
it's
just much less likely to compress a delta now (since they're small).
I've
made WT-15032 to cover this.

It also doesn't break any of our tests -- I was sure we had something
for
this, but I might be thinking of the server's delta testing code that
has a
compression scenario.