WT-13074 Fix log_newfile related warnings from ex_backup (#12122)
The pull request aims to fix the log_newfile related warinings from
ex_backup. The only concurrent operations that can touch shared
variables is between an application thread and background threads.
Applications threads that need to access to the same slot and wants to
change the state of the slot needs exclusive access using
WTI_WITH_SLOT_LOCK. I have posted the relevant TSAN warnings inside the
ticket.

(1) Inside log_newfile function we wait for the log_close_fh to be NULL.
The log_file_server is conditionally signalled for the log_close_fh to
be NULL. It is the only concurrency that can happen as log_newfile
requires a slot lock. Thus an atomic relax is enough.

(2) The **log->log_close_lsn** needs to be updated correct related to
**log->log_close_fh**. In log_newfile perform an WT_RELEASE_WITH_BARRIER
to ensure that **log->log_close_lsn** must be written before
log->log_close_fh is set. As the same time, we ensure this behaviour
with an **WT_ACQUIRE_READ_WITH_BARRIER** inside log_file_server.
Therefore an relaxed atomic is okay here.

(3) The **__log_prealloc_once** function is only called from the
background **log_server** thread. It is possible that it contends with
an application thread calling **__log_newfile**. This contention is not
necessary as the **__log_newfile** only checks if the configuration has
been set or not. I have used prealloc_init_count to acommodate for this
which is non-changing.

**Note: There has been no critical bug in log system for over 5 years**