WT-13746 Don't leave dirty btrees after RTS/recovery/shutdown (8.0) (#12133)
The changes ensure WiredTiger does not leave btrees dirty in specific
scenarios such as RTS, recovery and shutdown. In those scenarios, since
there should not be other actors using the btrees, checkpoint should be
able to process them and mark them clean.

Note that changes from WT-13618 are also added in this PR:
The `rec_max_timestamp` field inside the btree structure tracks the
maximum timestamp seen by reconciliation. This field was not updated
correctly in case of a prepared transaction that is rolled back. The
changes ensure this field is correctly updated.