WT-15106 Make cache walk heuristic reset specific to disaggregated storage (#12201)
WT-14813 introduced code that prioritised evicting pages with more dirty
content. Most of the changes were restricted to disaggregated storage,
but this change that reset the eviction walk flag heuristic was not, and
it seems to be responsible for a regression in performance.

Since this change was originally made for a disaggregated storage
specific change to dirty eviction heuristics, we should also make this
change specific to disaggregated storage.