WT-15114 Fix test_verify_disagg on MacOS (#12207)
Currnently `test_verify_disagg` is failing on macOS.

The reason is that the test sometimes expects errors to be reported and
checks whether the error message matches the expected one. However, the
expected messages are currently set manually, and I discovered that the
same error codes can produce slightly different messages on different
operating systems.

The fix is to use error numeric values + `strerror()` to retrieve the
message corresponding to the error code.