WT-14614 Improve error message when parsing unexpected escaped character (#12194)
The newline character enclosed in quotes is currently treated as an
unexpected character by the config string parsing function in
wiredtiger. The error message will be broken down into 2 lines when
printed out (thanks to the newline character).

The change in this PR tries to improve the error message by mentioning
it's an escape character for newline '\n', carriage return '\r', tab
'\t', backspace '\b', all of which are treated as unexpected characters
within quotes.