WT-14875 Create a macro to check default session (#12208)
In the current codebase, a few places have the check of `session->id`
value against 0 as a measure to determine whether it's a default session
or not. It's not intuitive to understand and usually requires an
additional comment to describe the intention.

This PR introduces a new macro to perform the default session check and
adopts the macro in various places where `session->id` value is checked
against 0.