WT-14941 Write prepared updates to disk for checkpoint when the preserve prepared config is enabled (#12156)
This is the initial version that implements the writing a complete
prepared transaction to disk when the preserve prepare config is
enabled. There are a lot of edge cases it handles:

- [x] Skip an aborted prepared update if its rollback timestamp is
stable.
- [x] Skip an aborted prepared update if its prepared timestamp is not
stable.
- [x] Writing an aborted prepared update to disk.
- [x] Skip a committed prepared update if its prepared timestamp is not
stable and leave the page dirty.
- [x] Write a committed prepared update as a prepared update if its
prepared timestamp is stable but its durable timestamp is not stable.
Leave the page dirty.
- [x] Write a committed prepared update as a committed update if its
durable timestamp is stable. Leave the page clean.
- [x] Skip write a prepared update if its prepared timestamp is not
stable and leave the page dirty.
- [x] Write a prepared update if its prepared timestamp is stable. Leave
the page dirty if it is checkpoint.
- [x] Writing a prepared tombstone followed by an aborted prepared
update if we race with prepared rollback.
- [x] Reconstruct aborted prepared modifies. Handle the case that it
races with prepared rollback.
- [x] We may prepend a globally visible tombstone to delete the
rollbacked prepared updates. Ensure we also write the prepared update to
disk in this case.
- [ ] Handle memory ordering if we race with prepared rollback.
- [ ] We may restore the update chain with prepared update (prepared,
already resolved, or aborted) to the new disk image after writing a
prepared update to disk. When instantiating the new disk image, we also
reinstantiate a new prepared update to the update chain. We now have
multiple prepared updates (some are resolved, aborted, some are still
prepared) on the update chain. This is very nasty.

---------

Co-authored-by: Linh Tran <linh.tran@mongodb.com>