WT-15158 Avoid setting ingest_btree->prune_timestamp during initialization (#12242)
Currently during the ingest table initialization we set
`prune_timestamp` to the last known checkpoint timestamp. According to
the current logic it could cause the following assertion failure:
```c
__layered_update_gc_ingest_tables_prune_timestamps(...) {
     WT_ASSERT(session, prune_timestamp >= btree->prune_timestamp);
``` 

It happens because we always update timestamp to the last checkpoint's
timestamp for this table which is in use and it could be newer than the
last known checkpoint timestamp.

This logic does require a redesign which is the main scope for WT-15192
Since this issue blocks other activities it was decided to merge it
without a regression test, the regression test would be provided as a
separate PR as well (WT-15191)