WT-15095 Refine eviction selection of pages with few updates (#12255)
This is an update to a change made in WT-14813 that changed eviction in
disaggregated storage to skip queuing pages with fewer updates if there
is not much cache pressure.

The original change considered the cache to be low pressure if either
the update or dirty thresholds haven't been reached. This addresses a
perf-change and modifies it to consider the cache to be under pressure
if either one of update or dirty pages are close to the threshold rather
than both, so pages are queued.