WT-15191 Regression test for WT-15158 (#12260)
This PR contains a regression test for the issue related to pruning
timestamps for ingest table.

The logic before fix was initializing the `prune_timestamp` to the last
checkpoint timestamp while initializing an ingest table. However, during
a checkpoint pickup, we update `prune_timestamp` to the oldest
checkpoint which is currently in use.

There is a requirement that we should update timestamp only to newer
one.

So if we initialize `prune_timestamp` to `ckpt2` timestamp while
initializing, and then, during the checkpoint pickup, `ckpt1` would
still be inuse, it could cause rewriting the `ckpt2` timestamp by the
`ckpt1` timestamp which breaks the ordering rule.

The test in this PR reproduces this case.