WT-15168 Understand (but ignore) old configuration strings from metadata (#12256)
This is pretty dumb right now.

I originally had some code to check that the values for these
configurations keys was empty, but the motivating problem was actually
old configs. That's more work to check because it's on the open path. A
check on session->create is probably acceptable, but I'm not sure that's
enough coverage to bother with. I'd like some feedback/discussion on
this decision.

Finally, I also sneaked in a couple of `self.assertEqual` spelling fixes
for newer Python versions.