WT-15281 Fix prepare rollback lead to an update missed in delta (#12293)
In disagg, when building a delta with prepared updates, the following
situation can cause us missing an update in the delta:

* Initially we write update U to the data store and we mark it with the
WT_UPDATE_DURABLE flag.
* We do a prepared update Up. Now we have Up -> U on the update chain.
* We rollback the prepared update.
* However, the rollback timestamp is not stable and we still write Up to
the data store and U to the history store.
* The rollback timestamp becomes stable and this time we skip Up and
skip U as well when building the delta because it already has the
WT_UPDATE_DURABLE flag on it.

We know that we need to delete U from the history store in this case as
we will write it to the data store. We can clear the WT_UPDATE_DURABLE
when saving the update to be deleted from the history store.