WT-15238 Fix prepared commit and rollback race with reconciliation (#12276)
Change the way we handle prepared commit and rollback. It will no longer
mark anything to be deleted from the history store. It will just append
the history store record to the update chain if necessary and resolve
the prepared updates. Reconciliation will handling deleting the updates
from the history store and inserting updates to the history store alone.
It is doing that based on the new flag WT_UPDATE_HS_MAX_STOP, which
means an update is inserted into the history store with a max time
point. Reconciliation will need to delete the update with this flag from
the history store if we write a committed update to the data store. In
addition, if we select an update has been inserted to the history store
to write to the data store, we also need to delete that record from the
history store first.

Fix an issue that update restore eviction for prepared updates need to
retain the oldest full value to resolve modifies. Add a test for that.

Fix a race that checkpoint may see an aborted update in prepare locked
state.

Fix a case that prepare insert rollback is lost. The scenario is like
this:
* We have a prepared insert.
* We rollback the prepared insert.
* Eviction writes it as prepared to disk.
* We restore back the aborted prepared.
* The rollback timestamp becomes stable.
* We ignore it and still keep the existing on-disk prepared as it is.
* Another reader reads the page and hits a prepared conflict.