WT-15299 Use checkpoint's pinned id for visibility in eviction for precise checkpoint (#12314)
The previous code only considers the checkpoint id for eviction.
However, what really matters is the checkpoint's snapshot. When precise
checkpoint is enabled, ideally we should use checkpoint's snapshot for
the eviction threads. However, that is difficult because of the
checkpoint can concurrently release its snapshot. Therefore, we pin the
checkpoint's pinned id instead.

In addition, we allow eviction to use the newest snapshot to evict if
checkpoint has visited the tree to alleviate some cache pressure.