WT-15165 POC Use a hashmap for prepare discover cursor (#12297)
This is a POC for replacing the array-based pending_prepared_sessions
with a hash map structure
for better scalability and memory efficiency. This change:

- Introduces WT_PENDING_PREPARED_ITEM and WT_PENDING_PREPARED_MAP
structures
- Refactors transaction claiming to use direct memory transfer instead
of session swapping
- Adds hash-based lookup for prepared transaction discovery

TODOs (will do these in other PRs):
- Add locks to the hash map. Initial prototype says there will be no
concurrency issues with this path, so I made the map single-threaded.
But there might be a problem when multiple sessions trying to access the
same bucket in the map
- Remove item from the map once prepared txn is claimed