WT-15086 Fix eviction thread panic during connection close with disagg enabled (#12307)
Creating a checkpoint in WiredTiger with attached storage results in
creating a checkpoint, even if the only change to the database was
advancing the stable timestamp—even though there were no changes to the
data. This PR brings the same semantics to WiredTiger for diaggregated
storage to result in having consistent behavior.

Also, add an error message if anyone attempts to create a precise
checkpoint without a stable timestamp.