WT-15210 Add scrub eviction feature flag changes (#12302)
This uses the new eviction_control feature flags and adds a toggle that
allows scrub eviction as long as we are less than half way to the clean
cache trigger. It no longer looks at dirty and update triggers as
metrics to allow or prevent the retention of clean pages in cache.