WT-14836 Add unclean shutdown to compatibility tests (#12274)
This adds a new `-d` (for "dirty") flag to the compatibility test
script.

It does this by first running `test/format` with the `format.abort`
flag. Once this is done (i.e. has crashed) we run a newer`test/format`
on the files the first run generated.

Thanks Sid for validating the original idea and making sure the first
run does what we thought it should!